+++ platform-paper/src/main/kotlin/dev/m1sk9/lunaticChat/paper/command/impl/lc/channel/ChannelInfoCommand.kt
@@ -33,6 +33,9 @@ class ChannelInfoCommand(
         return applyMethodPermission("build", builder)
     }
 
+    fun buildAllWithPermissionCheck(): List<LiteralArgumentBuilder<CommandSourceStack>> =
+        withAliases(buildWithPermissionCheck(), listOf("i"))
+
     @Permission(LunaticChatPermissionNode.ChannelInfo::class)
     fun build(): LiteralArgumentBuilder<CommandSourceStack> =
         Commands