+++ platform-paper/src/main/kotlin/dev/m1sk9/lunaticChat/paper/command/impl/lc/channel/ChannelListCommand.kt
@@ -35,6 +35,9 @@ class ChannelListCommand(
         return applyMethodPermission("build", builder)
     }
 
+    fun buildAllWithPermissionCheck(): List<LiteralArgumentBuilder<CommandSourceStack>> =
+        withAliases(buildWithPermissionCheck(), listOf("ls"))
+
     @Permission(LunaticChatPermissionNode.ChannelList::class)
     fun build(): LiteralArgumentBuilder<CommandSourceStack> =
         Commands