+++ platform-paper/src/main/kotlin/dev/m1sk9/lunaticChat/paper/command/impl/lc/channel/ChannelOwnershipCommand.kt
@@ -31,6 +31,9 @@ class ChannelOwnershipCommand(
         return applyMethodPermission("build", builder)
     }
 
+    fun buildAllWithPermissionCheck(): List<LiteralArgumentBuilder<CommandSourceStack>> =
+        withAliases(buildWithPermissionCheck(), listOf("own"))
+
     @Permission(LunaticChatPermissionNode.ChannelOwnership::class)
     fun build(): LiteralArgumentBuilder<CommandSourceStack> =
         Commands